Get jTRACE lexicon

Get jTRACE lexicon

jtrace_get_lexicon(lexicon = NULL)

Arguments

lexicon
Character vector of length 1 indicating the jTRACE lexicon to import

Value

A data frame with the contents of the retrieved jTRACE lexicon. This data frame will always contain a column for word forms (phonology). Sometimes, it will also contain another column for lexical frequencies (frquency), depending on the information provided in that file in the jTRACE original implementation.

Examples

jtrace_get_lexicon("sevenlex")

#>        phonology frequency
#> 1     xgrixbl           NA
#> 2     xdxltri           NA
#> 3     xblikli           NA
#> 4     tartxri           NA
#> 5     talrxbl           NA
#> 6     sxbgrup           NA
#> 7     suprsid           NA
#> 8     starbrd           NA
#> 9     Srxbxri           NA
#> 10    skxlptr           NA
#> 11    skulgrl           NA
#> 12    skarlxt           NA
#> 13    salxdli           NA
#> 14    prxdusr           NA
#> 15    priskul           NA
#> 16    priistr           NA
#> 17    pradxkt           NA
#> 18    larkspr           NA
#> 19    lablali           NA
#> 20    kxrakxs           NA
#> 21    krusxbl           NA
#> 22    brxskli           NA
#> 23    brakxli           NA
#> 24    blxdSat           NA
#> 25    blxdrut           NA
#> 26    blukalr           NA
#> 27    batista           NA
#> 28    bakskar           NA
#> 29    akskart           NA