Phoneme inventory

A dataset containing an extensive set of phonemes that can be transcribed from IPA or SAMPA to jTRACE notation using approximations. These dimensions are the same used in jTRACE (Strauss et al. 2007) .

data(phonemes)

data("phonemes")

Format

A data frame with 74 rows and 7 variables:

id

Individual phoneme numeric identifier

ipa

Phoneme symbol in IPA notation

trace

Phoneme symbol in jTRACE notation

description

Rough classification of the phoneme

is_english

Is this phoneme present in English?

is_spanish

Is this phoneme present in Spanish?

is_catalan

Is this phoneme present in Catalan?

type

Is the phoneme a consonant or a vowel?

pow

Power dimension (McClelland and Elman 1986)

voc

Vocalic dimension (Jakobson et al. 1951)

dif

Diffusiveness dimension (Jakobson et al. 1951)

acu

Acuteness dimension (Jakobson et al. 1951)

con

Consonantal dimension (Jakobson et al. 1951)

voi

Voicing dimension (Jakobson et al. 1951)

bur

Burst dimension (McClelland and Elman 1986)

References

Jakobson R, Fant CG, Halle M (1951). Preliminaries to speech analysis: The distinctive features and their correlates. MIT press.

McClelland JL, Elman JL (1986). “The TRACE model of speech perception.” Cognitive psychology, 18(1), 1--86.

Strauss TJ, Harris HD, Magnuson JS (2007). “jTRACE: A reimplementation and extension of the TRACE model of speech perception and spoken word recognition.” Behavior Research Methods, 39(1), 19--30.

Examples

data("phonemes")